+++ src/components/home/ProductList/index.tsx
@@ -11,26 +11,29 @@ const ProductList = ({
   fetchNextPage
 }: ProductListProps): ReactElement => {
   const [isFirstRender, setIsFirstRender] = useState<boolean>(false)
-
   const { ref: isLastPrdRef, inView } = useInView({
     threshold: 1
   })
 
+  const shouldFetchNextPage = inView && hasNextPage
+
   useEffect(() => {
-    if (inView && hasNextPage) {
-      if (!isFirstRender) {
-        setIsFirstRender(true)
-        return
-      }
-      fetchNextPage && fetchNextPage()
+    if (!shouldFetchNextPage) {
+      return
     }
-  }, [inView, hasNextPage, fetchNextPage])
+    if (!isFirstRender) {
+      setIsFirstRender(true)
+      return
+    }
+
+    fetchNextPage && fetchNextPage()
+  }, [fetchNextPage, shouldFetchNextPage, isFirstRender])
 
   return (
     <>
       <Styled.NewProductTitle>새로운 상품</Styled.NewProductTitle>
       <Styled.ProductListWrapper>
-        {postData?.pages?.map(page =>
+        {postData?.map(page =>
           page?.posts?.map(item => (
             <ProductItem key={item.id} productItem={item} />
           ))
